+++ src/ExamplesProcessor.ts
@@ -18,23 +18,23 @@ export class ExamplesProcessor extends ListProcessor<Examples, ScenarioOutline> 
         this.tableRowProcessor = new TableRowProcessor<Examples>(preCompiler);
     }
 
-    protected preFilter(e: Examples, p: ScenarioOutline, i: number): boolean {
+    protected async preFilter(e: Examples, p: ScenarioOutline, i: number): Promise<boolean> {
         /* istanbul ignore next */
         debug(
             "preFilter(hasPreExamples: %s, e: %s, p: %s, i: %d)",
             !!this.preCompiler.preExamples, e?.constructor.name, p?.constructor.name, i
         );
-        return !this.preCompiler.preExamples || this.preCompiler.preExamples(e, p, i);
+        return !this.preCompiler.preExamples || await this.preCompiler.preExamples(e, p, i);
     }
-    protected postFilter(e: Examples, p: ScenarioOutline, i: number): boolean {
+    protected async postFilter(e: Examples, p: ScenarioOutline, i: number): Promise<boolean> {
         /* istanbul ignore next */
         debug(
             "postFilter(hasPostExamples: %s, e: %s, p: %s, i: %d)",
             !!this.preCompiler.postExamples, e?.constructor.name, p?.constructor.name, i
         );
-        return !this.preCompiler.postExamples || this.preCompiler.postExamples(e, p, i);
+        return !this.preCompiler.postExamples || await this.preCompiler.postExamples(e, p, i);
     }
-    protected process(e: Examples, p: ScenarioOutline, i: number): MultiControlType<Examples> {
+    protected async process(e: Examples, p: ScenarioOutline, i: number): Promise<MultiControlType<Examples>> {
         /* istanbul ignore next */
         debug(
             "process(hasOnExamples: %s, e: %s, p: %s, i: %d)",
@@ -42,7 +42,7 @@ export class ExamplesProcessor extends ListProcessor<Examples, ScenarioOutline> 
         );
         let examples: MultiControlType<Examples> = e;
         if (this.preCompiler.onExamples) {
-            const result = this.preCompiler.onExamples(e, p, i);
+            const result = await this.preCompiler.onExamples(e, p, i);
             if (typeof result !== "undefined") {
                 examples = result;
             }
@@ -50,20 +50,20 @@ export class ExamplesProcessor extends ListProcessor<Examples, ScenarioOutline> 
         if (examples) {
             if (Array.isArray(examples)) {
                 debug("...Array: %d", examples.length);
-                examples = examples.map(this.postProcess.bind(this));
+                examples = await ExamplesProcessor.map(examples, this.postProcess.bind(this));
             } else {
                 debug("...replace");
-                examples = this.postProcess(examples);
+                examples = await this.postProcess(examples);
             }
         }
         return examples;
     }
 
-    private postProcess(e: Examples): Examples {
+    private async postProcess(e: Examples): Promise<Examples> {
         debug("postProcess(tags: %s, body: %s)", Array.isArray(e.tags), Array.isArray(e.body));
-        e.tags = this.tagsProcessor.execute(e.tags, e);
-        e.header = this.tableRowProcessor.execute([e.header], e)?.[0];
-        e.body = this.tableRowProcessor.execute(e.body, e);
+        e.tags = await this.tagsProcessor.execute(e.tags, e);
+        e.header = (await this.tableRowProcessor.execute([e.header], e))?.[0];
+        e.body = await this.tableRowProcessor.execute(e.body, e);
         return e;
     }
 }