+++ src/cli/module/precompile.ts
@@ -1,11 +1,14 @@
+import {hasMagic, sync} from "glob";
 import type {Argv} from "yargs";
 import {dirname, join, normalize, resolve} from "path";
 import * as fs from "fs";
+import {getDebugger} from "./debug";
+import {PreCompiler} from "./PreCompiler";
+import {Document, ParseConfig, TagFormat} from "gherkin-ast";
 import {FormatOptions} from "gherkin-io";
-import {hasMagic, sync} from "glob";
-import {getDebugger} from "../../debug";
-import {PreCompiler} from "../../PreCompiler";
-import {Document} from "gherkin-ast";
+import yargs = require("yargs/yargs");
+// @ts-ignore
+import lazy = require("lazy-require");
 import {load, process as processAst, save} from "../../index";
 
 const debug = getDebugger("cli:precompile");
@@ -33,14 +36,19 @@ export interface CLIConfig {
     destination?: string;
     verbose?: boolean;
     clean?: boolean;
+    install?: boolean;
 }
 
-interface Config extends CLIConfig {
+export interface Config extends CLIConfig {
     compilers: CompilerConfig[];
     formatOptions?: FormatOptions;
+    parseConfig?: ParseConfig;
 }
 
 const isPackage = (name: string): boolean => {
+    if (/^gpc-[a-z-]+$/.test(name)) {
+        return true;
+    }
     try {
         require(name);
         return true;
@@ -63,6 +71,67 @@ const isFile = (path: string): boolean => {
     } catch (e) {
         return false;
     }
+}
+
+const resolvePath = (path: string): string => path ? resolve(path) : null;
+
+const parseConfig = (): Config => {
+    debug("parseConfig %o", process.argv);
+    return yargs(process.argv)
+        .usage("Usage: gherking --config <path> [options]")
+        .option("config", {
+            type: "string",
+            alias: "c",
+            coerce: resolvePath,
+            default: "./.gherking.json",
+            description: "The path of the configuration file which contains the precompilers and their configurations.",
+            normalize: true,
+            config: true,
+            configParser: path => require(path)
+        })
+        .option("source", {
+            type: "string",
+            alias: "s",
+            description: "The pattern or path of feature files which needs to be precompiled.",
+            normalize: true
+        })
+        .option("base", {
+            type: "string",
+            alias: "b",
+            description: "The base directory of feature files.",
+            normalize: true
+        })
+        .option("destination", {
+            type: "string",
+            alias: "d",
+            description: "The destination directory of precompiled feature files.",
+            normalize: true
+        })
+        .option("install", {
+            type: "boolean",
+            description: "Whether the missing precompilers (gpc-* packages) should be installed and save to the package.json. "
+                + "Packages will be installed in the current folder, and package.json created if it is not there yet.",
+            default: false,
+        })
+        .option("verbose", {
+            type: "boolean",
+            description: "Whether some information should be displayed on the screen.",
+            default: false,
+        })
+        .option("clean", {
+            type: "boolean",
+            description: "Whether the destination directory should be clean in advance.",
+            default: false,
+        })
+        .check(argv => prepareConfig(argv as unknown as Config))
+        .help("help")
+        .locale("en")
+        .fail((msg, err, ya) => {
+            console.error(msg);
+            console.error(ya.help())
+            if (err) throw err;
+        })
+        .argv as unknown as Config;
 }
 
 const prepareConfig = (argv: Config): Config => {
@@ -118,9 +187,39 @@ const prepareConfig = (argv: Config): Config => {
     if (argv.verbose) {
         console.log("Configuration:", JSON.stringify(argv, null, 2));
     }
+    if (!argv.parseConfig)  {
+        // @ts-ignore
+        argv.parseConfig = {};
+    } else if (typeof argv.parseConfig.tagFormat === "string") {
+        const format = (argv.parseConfig.tagFormat as string).toUpperCase();
+        if (!(format in TagFormat)) {
+            throw new Error(`Input tag format is not supported: ${format}!`);
+        }
+        // @ts-ignore
+        argv.parseConfig.tagFormat = TagFormat[format];
+    }
+    if (!argv.formatOptions) {
+        argv.formatOptions = {};
+    } else if (typeof argv.formatOptions.tagFormat === "string") {
+        const format = (argv.formatOptions.tagFormat as string).toUpperCase();
+        if (!(format in TagFormat)) {
+            throw new Error(`Output tag format is not supported: ${format}!`);
+        }
+        // @ts-ignore
+        argv.formatOptions.tagFormat = TagFormat[format];
+    }
+    if (!argv.formatOptions.tagFormat) {
+        argv.formatOptions.tagFormat = argv.parseConfig.tagFormat;
+    }
     return argv;
 }
 
+export interface LoadOptions {
+    install?: boolean;
+}
+
+const loadCompilers = (compilers: CompilerConfig[], options: LoadOptions = {}): PreCompiler[] => {
+    debug("loadCompilers(compilers: %d, options: %o)", compilers.length, options);
 const resolvePath = (path: string): string => path ? resolve(path) : null;
 
 export function builder(yargs: Argv) {
@@ -175,7 +274,11 @@ const loadCompilers = (compilers: CompilerConfig[]): PreCompiler[] => {
     return compilers.map(compiler => {
         let preCompiler;
         if (isPackage(compiler.path)) {
-            preCompiler = require(compiler.path);
+            if (options?.install) {
+                preCompiler = lazy(compiler.path);
+            } else {
+                preCompiler = require(compiler.path);
+            }
         } else {
             preCompiler = require(resolve(compiler.path));
         }
@@ -195,48 +298,62 @@ const loadCompilers = (compilers: CompilerConfig[]): PreCompiler[] => {
     });
 };
 
-const processSource = async (source: IOConfig, compilers: PreCompiler[], formatOptions: FormatOptions): Promise<void> => {
+const getSources = (config: Config): IOConfig[] => {
+    debug("getSources(source: %s, base: %s, dest: %s)", config.source, config.base, config.destination);
+    return sync(config.source.replace(/\\/g, "/"))
+        .map(normalize)
+        .map(file => {
+            debug("getSources:source: %s", file)
+            return ({
+                input: file,
+                output: join(config.destination, file.replace(config.base, "")),
+            });
+        });
+}
+
+const processSource = async (source: IOConfig, compilers: PreCompiler[], parseConfig: ParseConfig, formatOptions: FormatOptions): Promise<void> => {
     const outputDir = dirname(source.output);
     if (!fs.existsSync(outputDir)) {
         fs.mkdirSync(outputDir, { recursive: true });
     }
 
-    const documents: Document[] = await load(source.input);
+    const documents: Document[] = await load(source.input, parseConfig);
     const document = documents[0];
     document.targetFolder = outputDir;
-    const outputAst = processAst(document, ...compilers);
+    const outputAst = await processAst(document, ...compilers);
     await save(source.output, outputAst, formatOptions);
-}
-
-const getSources = (config: Config): IOConfig[] => {
-    return sync(config.source, { absolute: true })
-        .map(normalize)
-        .map(file => ({
-            input: file,
-            output: config.destination + file.replace(config.base, ""),
-        }));
 }
 
 export async function handler(argv: Config): Promise<void> {
     debug("run");
-    const config = prepareConfig(argv);
-    debug("...config: %o", config);
+    const config = parseConfig();
+    debug("...config: %s", JSON.stringify(config));
 
-    const compilers = loadCompilers(config.compilers);
+    lazy._installSync = lazy.installSync;
+    /* istanbul ignore next */
+    lazy.installSync = function (...args: unknown[]): void {
+        debug("...installSync: %o", args);
+        config.verbose && console.log(`Installing ${args[0]}`);
+        return lazy._installSync(...args);
+    };
+
+    const compilers = loadCompilers(config.compilers, config);
     debug("...compilers: %o", compilers);
 
     const sources = getSources(config);
     debug("...sources: %o", sources);
 
     if (config.clean && fs.existsSync(config.destination)) {
+        debug("...clean: %s", config.destination);
         config.verbose && console.log(`Cleaning ${config.destination}`);
         (fs.rmSync ? fs.rmSync : fs.rmdirSync)(config.destination, { recursive: true });
         fs.mkdirSync(config.destination, { recursive: true });
     }
 
     for (const source of sources) {
+        debug("...processing: %o", source);
         config.verbose && console.log(`Processing ${source.input}`);
-        await processSource(source, compilers, config.formatOptions);
+        await processSource(source, compilers, config.parseConfig, config.formatOptions);
         config.verbose && console.log(`Processed file written out ${source.output}`);
     }
 }