+++ src/index.ts
@@ -1,6 +1,6 @@
 import { PreCompiler } from "./PreCompiler";
 import { read, write, FormatOptions } from "gherkin-io";
-import { Document } from "gherkin-ast";
+import { Document, ParseConfig } from "gherkin-ast";
 import { DocumentProcessor } from "./DocumentProcessor";
 import { getDebugger } from "./debug";
 
@@ -10,13 +10,13 @@ export * from "./PreCompiler";
 export * from "gherkin-ast";
 export { FormatOptions } from "gherkin-io";
 
-export const load = (pattern: string): Promise<Document[]> => {
-    debug("load(pattern: %s)", pattern);
+export const load = (pattern: string, config?: ParseConfig): Promise<Document[]> => {
+    debug("load(pattern: %s, config: %o)", pattern, config);
     // @ts-ignore
-    return read(pattern) as Document[];
+    return read(pattern, config) as Document[];
 }
 
-export const process = (ast: Document, ...preCompilers: PreCompiler[]): Document[] => {
+export const process = async (ast: Document, ...preCompilers: PreCompiler[]): Promise<Document[]> => {
     /* istanbul ignore next */
     debug("process(ast: %s, preCompilers: %d)", ast?.constructor.name, preCompilers.length);
     const documents = [ast];
@@ -25,7 +25,7 @@ export const process = (ast: Document, ...preCompilers: PreCompiler[]): Document
         const processor = new DocumentProcessor(preCompiler);
         for (let i = 0; i < documents.length; ++i) {
             debug("......document: %d", i);
-            const newDocuments = processor.execute(documents[i]);
+            const newDocuments = await processor.execute(documents[i]);
             debug("......new documents: %d", newDocuments.length);
             documents.splice(i, 1, ...newDocuments);
             i += newDocuments.length - 1;