+++ tests/RuleProcessor.test.ts
@@ -19,17 +19,17 @@ describe("RuleProcessor", ()=> {
          ];
     })
     
-    test("should handle if no pre/post-filter or event handler is set", () => {
+    test("should handle if no pre/post-filter or event handler is set", async () => {
         const items1 = feature1.elements;
         const items2 = feature2.elements;
         const ruleProcessor = new RuleProcessor();
-        const result1 = ruleProcessor.execute(items1 as Rule[], feature1);
-        const result2 = ruleProcessor.execute(items2 as Rule[], feature2);
+        const result1 = await ruleProcessor.execute(items1 as Rule[], feature1);
+        const result2 = await ruleProcessor.execute(items2 as Rule[], feature2);
         expect(result1).toEqual([rule1]);
         expect(result2).toEqual([rule1, rule2]);
     });
 
-    test("should filter by pre-filter", () => {
+    test("should filter by pre-filter", async () => {
         const items1 = feature1.elements;
         const items2 = feature2.elements;
         const onRule = jest.fn();
@@ -39,13 +39,13 @@ describe("RuleProcessor", ()=> {
             },
             onRule
         });
-        const result1 = ruleProcessor.execute(items1 as Rule[], feature1);
-        const result2 = ruleProcessor.execute(items2 as Rule[], feature2);
+        const result1 = await ruleProcessor.execute(items1 as Rule[], feature1);
+        const result2 = await ruleProcessor.execute(items2 as Rule[], feature2);
         expect(result1).toEqual([]);
         expect(result2).toEqual([rule2]);
     });
     
-    test("should filter by post-filter", () => {
+    test("should filter by post-filter", async () => {
         const items1 = feature1.elements;
         const items2 = feature2.elements;
         const onRule = jest.fn();
@@ -55,13 +55,13 @@ describe("RuleProcessor", ()=> {
             },
             onRule
         });
-        const result1 = ruleProcessor.execute(items1 as Rule[], feature1);
-        const result2 = ruleProcessor.execute(items2 as Rule[], feature2);
+        const result1 = await ruleProcessor.execute(items1 as Rule[], feature1);
+        const result2 = await ruleProcessor.execute(items2 as Rule[], feature2);
         expect(result1).toEqual([rule1]);
         expect(result2).toEqual([rule1]);
     });
 
-    test("should process array of rules with event handler", () => {
+    test("should process array of rules with event handler", async () => {
         const items = feature2.elements;
         const ruleProcessor = new RuleProcessor({
             onRule(e: Rule): Array<Rule> {
@@ -70,7 +70,7 @@ describe("RuleProcessor", ()=> {
                 return [c, c];
             },
         });
-        const results = ruleProcessor.execute(items as Rule[], feature2);
+        const results = await ruleProcessor.execute(items as Rule[], feature2);
 
         const names = results.map(s => s.name)
 
@@ -78,7 +78,7 @@ describe("RuleProcessor", ()=> {
         expect(names).toEqual(["21", "21", "51", "51"]);
     });
 
-    test("should process scenarios", () => {
+    test("should process scenarios", async () => {
         rule1.elements = [new Scenario("Scenario", "1", "")];
         const ruleProcessor = new RuleProcessor({
             onRule(e: Rule) {
@@ -88,7 +88,7 @@ describe("RuleProcessor", ()=> {
                 e.name += "PROCESSED";
             }
         });
-        const results = ruleProcessor.execute(feature1.elements as Rule[], feature1);
+        const results = await ruleProcessor.execute(feature1.elements as Rule[], feature1);
         expect(results[0].name).toContain("PROCESSED");
         expect(results[0].elements[0].name).toContain("PROCESSED");
     });