+++ tests/TableRowProcessor.test.ts
@@ -19,14 +19,14 @@ describe("TableRowProcessor", () => {
         dataTable = new DataTable(rows);
     });
 
-    test("should handle if no pre/post-filter or event handler is set", () => {
+    test("should handle if no pre/post-filter or event handler is set", async () => {
         const tableRowProcessor = new TableRowProcessor<DataTable>();
-        const results = tableRowProcessor.execute(rows, dataTable);
+        const results = await tableRowProcessor.execute(rows, dataTable);
 
         expect(results).toEqual(rows);
     });
 
-    test("should filter by pre-filter", () => {
+    test("should filter by pre-filter", async () => {
         const tableRowProcessor = new TableRowProcessor<DataTable>({
             preTableRow(e: TableRow): boolean {
                 return e.cells.some(c => c.value === "1");
@@ -35,7 +35,7 @@ describe("TableRowProcessor", () => {
                 e.cells.push(new TableCell("new"));
             },
         });
-        const results = tableRowProcessor.execute(rows, dataTable);
+        const results = await tableRowProcessor.execute(rows, dataTable);
         expect(results).toHaveLength(1);
         expect(results[0].cells).toHaveLength(3);
         expect(results[0].cells[0].value).toBe("1");
@@ -43,7 +43,7 @@ describe("TableRowProcessor", () => {
         expect(results[0].cells[2].value).toBe("new");
     });
 
-    test("should filter by post-filter", () => {
+    test("should filter by post-filter", async () => {
         const tableRowProcessor = new TableRowProcessor<DataTable>({
             postTableRow(e: TableRow): boolean {
                 return e.cells.some(c => c.value === "5");
@@ -54,7 +54,7 @@ describe("TableRowProcessor", () => {
                 ));
             },
         });
-        const results = tableRowProcessor.execute(rows, dataTable);
+        const results = await tableRowProcessor.execute(rows, dataTable);
         expect(results).toHaveLength(1);
         expect(results[0].cells).toHaveLength(3);
         expect(results[0].cells[0].value).toBe("3");
@@ -62,13 +62,13 @@ describe("TableRowProcessor", () => {
         expect(results[0].cells[2].value).toBe("5");
     });
 
-    test("should process with event handler", () => {
+    test("should process with event handler", async () => {
         const tableRowProcessor = new TableRowProcessor<DataTable>({
             onTableRow(e: TableRow): void {
                 e.cells.push(new TableCell("new"));
             },
         });
-        const results = tableRowProcessor.execute(rows, dataTable);
+        const results = await tableRowProcessor.execute(rows, dataTable);
 
         expect(results).toHaveLength(2);
         for (const row of results) {